+++ Source/Relativity.Test.Helpers/Relativity.Test.Helpers.NUnit.Integration/UserHelpers/UserHelpersIntegrationTests.cs
@@ -1,13 +1,9 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using kCura.Relativity.Client;
+﻿using kCura.Relativity.Client;
 using kCura.Relativity.Client.DTOs;
 using NUnit.Framework;
 using Relativity.API;
-using Relativity.Test.Helpers.SharedTestHelpers;
+using System;
+using System.Collections.Generic;
 using User = kCura.Relativity.Client.DTOs.User;
 
 namespace Relativity.Test.Helpers.NUnit.Integration.UserHelpers